+++ mlf_core/create/create.py
@@ -1,4 +1,5 @@
 from mlf_core.create.domains.mlflow_creator import MlflowCreator
+from mlf_core.create.domains.package_creator import PackageCreator
 from mlf_core.custom_cli.questionary import mlf_core_questionary_or_dot_mlf_core
 
 
@@ -14,13 +15,14 @@ def choose_domain(domain: str or None, dot_mlf_core: dict = None):
     if not domain:
         domain = mlf_core_questionary_or_dot_mlf_core(function='select',
                                                       question='Choose the project\'s domain',
-                                                      choices=['mlflow'],
+                                                      choices=['mlflow', 'package'],
                                                       default='mlflow',
                                                       dot_mlf_core=dot_mlf_core,
                                                       to_get_property='domain')
 
     switcher = {
-        'mlflow': MlflowCreator
+        'mlflow': MlflowCreator,
+        'package': PackageCreator
     }
 
     creator_obj = switcher.get(domain.lower())()