+++ mlf_core/create/template_creator.py
@@ -224,13 +224,8 @@ class TemplateCreator:
                                                          dot_mlf_core=dot_mlf_core,
                                                          to_get_property='version')
 
-        def suffix_check(version: str) -> bool:
-            if len(version) > 6 and '-SNAPSHOT' not in version:
-                return False
-            return True
-
         # make sure that the version has the right format
-        while not (re.match(r'(?<!\.)\d+(?:\.\d+){2}(-SNAPSHOT)?(?!\.)', poss_vers) and suffix_check(poss_vers)) and not dot_mlf_core:
+        while not re.match(r'(?<!.)\d+(?:\.\d+){2}(?:-SNAPSHOT)?(?!.)', poss_vers) and not dot_mlf_core:
             print('[bold red]The version number entered does not match semantic versioning.\n' +
                   'Please enter the version in the format \[number].\[number].\[number](-SNAPSHOT)!')  # noqa: W605
             poss_vers = mlf_core_questionary_or_dot_mlf_core(function='text',